% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concurrence.R
\name{adorn_common_percentages}
\alias{adorn_common_percentages}
\title{Adorn the concurrence table}
\usage{
adorn_common_percentages(M, envname = "env", digits = 1L)
}
\arguments{
\item{M}{The concurrence/connectivity matrix.}

\item{envname}{The name of the environment column.}

\item{digits}{An integer specifying the number of decimal places.}
}
\description{
This function is similar to the adorn_ functions in the janitor package.
It adorns the a frequency table with relevant statistics. In this case,
it returns a data.frame with the number of genotypes or common genotypes across environment.
The percentage shows the proportion of common genotypes across environments where the denominator is
the number of genotypes in the corresponding row environment.
}
