% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concurrence.R
\name{concurrence}
\alias{concurrence}
\alias{connectivity}
\title{Get the concurrence matrix}
\usage{
concurrence(data, gen, env, sort = "none")

connectivity(data, gen, env, sort = "none")
}
\arguments{
\item{data}{A data.}

\item{gen}{The name of the genotype column.}

\item{env}{The name of the environment column.}

\item{sort}{The sorting method. "freq" sorts by descending diagonal frequency.}
}
\description{
Computes the concurrence matrix for a given GxE data. The connectivity matrix
is a synonym for the concurrence matrix. The diagonal elements return the number
of genotypes in the corresponding environment. The off-diagonal elements are the
number of common genotypes across environments.
}
