% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{genotype_fuzzy_match_left}
\alias{genotype_fuzzy_match_left}
\alias{genotype_fuzzy_match_right}
\alias{genotype_fuzzy_match_full}
\title{Attempts to find a closely matching genotype}
\usage{
genotype_fuzzy_match_left(
  x,
  y,
  ...,
  name_x = "x",
  name_y = "y",
  ignore_case = TRUE
)

genotype_fuzzy_match_right(
  x,
  y,
  ...,
  name_x = "x",
  name_y = "y",
  ignore_case = TRUE
)

genotype_fuzzy_match_full(
  x,
  y,
  ...,
  name_x = "x",
  name_y = "y",
  ignore_case = TRUE
)
}
\arguments{
\item{x, y}{A pair of vectors that contain the genotype labels in a pair of datasets.}
}
\value{
A data.frame returning the genotype labels that are closely matched.
}
\description{
Attempts to find a closely matching genotype
}
