% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{genotype_mismatch_left}
\alias{genotype_mismatch_left}
\alias{genotype_mismatch_right}
\alias{genotype_mismatch_full}
\title{Genotype mismatch}
\usage{
genotype_mismatch_left(x, y)

genotype_mismatch_right(x, y)

genotype_mismatch_full(x, y)
}
\arguments{
\item{x, y}{A pair of vectors that contain the genotype labels in a pair of datasets.}
}
\value{
A vector of genotype labels that are mismatched.
}
\description{
Genotype mismatch
}
\seealso{
\code{\link[=genotype_fuzzy_match_left]{genotype_fuzzy_match_left()}}
}
